% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual-select.R
\name{virtualSelectInput}
\alias{virtualSelectInput}
\title{Virtual Select Input}
\usage{
virtualSelectInput(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  search = FALSE,
  hideClearButton = FALSE,
  showSelectedOptionsFirst = FALSE,
  showValueAsTags = FALSE,
  optionsCount = 10,
  ...,
  inline = FALSE,
  width = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of choices, \strong{warning} format is a little different than \code{\link[shiny:selectInput]{shiny::selectInput()}}.
You can use:
\itemize{
\item \code{vector} use a simple vector for better performance.
\item \code{list} custom list allowing to use more options, must correspond to \href{https://sa-si-dev.github.io/virtual-select/#/properties}{virtual-select specifications}
}}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). If not specified then defaults to the first value for
single-select lists and no values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{search}{Enable search feature.}

\item{hideClearButton}{Hide clear value button.}

\item{showSelectedOptionsFirst}{Show selected options at the top of the dropbox.}

\item{showValueAsTags}{Show each selected values as tags with remove icon.}

\item{optionsCount}{No.of options to show on viewport.}

\item{...}{Other arguments passed to JavaScript method, see
\href{https://sa-si-dev.github.io/virtual-select/#/properties}{virtual-select documentation} for a full list of options.}

\item{inline}{Display inline with label or not.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}
}
\value{
A \code{shiny.tag} object that can be used in a UI definition.
}
\description{
A select dropdown widget made for performance,
based on \href{https://github.com/sa-si-dev/virtual-select}{virtual-select}.
}
\examples{
library(shiny)
library(shinyvs)

ui <- fluidPage(
  tags$h2("Virtual Select"),

  virtualSelectInput(
    inputId = "single",
    label = "Single select (with 97310 choices) :",
    choices = sort(unique(babynames::babynames$name)),
    search = TRUE
  ),
  verbatimTextOutput("res_single"),

  virtualSelectInput(
    inputId = "multiple",
    label = "Multiple select:",
    choices = setNames(month.abb, month.name),
    selected = "Jun",
    multiple = TRUE
  ),
  verbatimTextOutput("res_multiple")
)

server <- function(input, output, session) {
  output$res_single <- renderPrint(input$single)
  output$res_multiple <- renderPrint(input$multiple)
}

if (interactive())
  shinyApp(ui, server)
}
